Decouple web_page and web_logger from error_prompt
