+++ src/browser/browser.py
@@ -78,11 +78,6 @@ import resources.rc_resources
 BridgeObjects = Tuple["BridgeObject"]
 Url = TypeVar("Url", str, QUrl)
 
-os.environ["QT_DEVICE_PIXEL_RATIO"] = "0"
-os.environ["QT_AUTO_SCREEN_SCALE_FACTOR"] = "1"
-os.environ["QT_SCREEN_SCALE_FACTORS"] = "1"
-os.environ["QT_SCALE_FACTOR"] = "1"
-
 
 def get_default_cursor():
     """Gets the default cursor theme"""
