+++ src/browser/error_prompt.py
@@ -35,118 +35,10 @@ from logging import (
     Formatter,
     StreamHandler,
 )
-from PySide6.QtWebEngineCore import QWebEnginePage
-from PySide6.QtWidgets import (
-    QAbstractButton,
-    QDialogButtonBox,
-    QDialog,
-    QVBoxLayout,
-    QLabel,
-    QPushButton,
-)
 from PySide6.QtGui import QWindow
 from config import web_greeter_config
 
 import globales
-
-LOG_FORMAT = "".join(
-    ["%(asctime)s [ %(levelname)s ] %(filename)s %(", "lineno)d: %(message)s"]
-)
-formatter = Formatter(fmt=LOG_FORMAT, datefmt="%Y-%m-%d %H:%M:%S")
-logger = getLogger("javascript")
-logger.propagate = False
-stream_handler = StreamHandler()
-stream_handler.setLevel(DEBUG)
-stream_handler.setFormatter(formatter)
-logger.setLevel(DEBUG)
-logger.addHandler(stream_handler)
-
-
-class WebPage(QWebEnginePage):
-    """web-greeter's webpage class"""
-
-    def javaScriptConsoleMessage(
-        self,
-        level: QWebEnginePage.JavaScriptConsoleMessageLevel,
-        message: str,
-        line_number: int,
-        source_id: str,
-    ):
-        # pylint: disable = no-self-use,missing-function-docstring,invalid-name
-        if source_id == "":
-            source_id = "console"
-
-        log_level = 0
-        if level == WebPage.ErrorMessageLevel:
-            log_level = 40
-        elif level == WebPage.WarningMessageLevel:
-            log_level = 30
-        elif level == WebPage.InfoMessageLevel:
-            return
-        else:
-            return
-
-        record = logger.makeRecord(
-            name="javascript",
-            level=log_level,
-            fn="",
-            lno=line_number,
-            msg=message,
-            args=(),
-            exc_info=None,
-        )
-        record.filename = source_id
-        logger.handle(record)
-
-        if log_level == 40:
-            errorMessage = f"{source_id} {line_number}: {message}"
-            # error_prompt(errorMessage)
-
-    def increaseZoom(self, value=0.1):
-        """Increase zoom"""
-        # pylint: disable=invalid-name
-        self.setZoomFactor(self.zoomFactor() + (value if value else 0.1))
-
-    def decreaseZoom(self, value=0.1):
-        """Increase zoom"""
-        # pylint: disable=invalid-name
-        self.setZoomFactor(self.zoomFactor() - (value if value else 0.1))
-
-
-class Dialog(QDialog):
-    """Popup dialog class"""
-
-    def __init__(
-        self,
-        parent=None,
-        title: str = "Dialog",
-        message: str = "Message",
-        detail: str = "",
-        buttons: List[str] = None,
-    ):
-        super().__init__(parent)
-        self.setWindowTitle(title)
-
-        self.button_box = QDialogButtonBox()
-        if buttons is not None:
-            for i, btn in enumerate(buttons, 0):
-                button = QPushButton(btn)
-                button.role = i
-                self.button_box.addButton(button, QDialogButtonBox.NoRole)
-
-        self.button_box.clicked.connect(self.handle_click)
-
-        self.layout = QVBoxLayout()
-        self.layout.addWidget(QLabel(message))
-        self.layout.addWidget(QLabel(detail))
-        self.layout.addWidget(self.button_box)
-
-        self.setLayout(self.layout)
-
-    def handle_click(self, button: QAbstractButton):
-        """Handle click of button"""
-        self.done(button.role)
-
 
 def general_error_prompt(window: QWindow, message: str, detail: str, title: str):
     """General error prompt"""
